+++ sacred/__about__.py
@@ -11,7 +11,7 @@ from __future__ import division, print_function, unicode_literals
 
 __all__ = ("__version__", "__author__", "__author_email__", "__url__")
 
-__version__ = "0.7b0"
+__version__ = "0.7b1"
 
 __author__ = 'Klaus Greff'
 __author_email__ = 'qwlouse@gmail.com'